# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.ErrorBody'
      meta: {}
      success:
        type: boolean
    type: object
  github_com_efangly_thanes-lims-backend_internal_adapters_http_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_adapters_http_inventory.CreateItemRequest:
    properties:
      category:
        type: string
      custodian_user_id:
        type: integer
      default_vendor:
        type: string
      location_id:
        type: string
      manufacturer:
        type: string
      max:
        type: integer
      min:
        type: integer
      name:
        type: string
      unit:
        type: string
      vendor_id:
        type: string
    required:
    - category
    - custodian_user_id
    - name
    - unit
    type: object
  internal_adapters_http_inventory.IssueLineRequest:
    properties:
      lot_id:
        type: string
      quantity:
        type: integer
    required:
    - lot_id
    type: object
  internal_adapters_http_inventory.IssueStockRequest:
    properties:
      force:
        type: boolean
      lines:
        items:
          $ref: '#/definitions/internal_adapters_http_inventory.IssueLineRequest'
        minItems: 1
        type: array
    required:
    - lines
    type: object
  internal_adapters_http_inventory.IssueStockResponse:
    properties:
      applied:
        type: boolean
      item:
        $ref: '#/definitions/internal_adapters_http_inventory.ItemResponse'
      lots:
        items:
          $ref: '#/definitions/internal_adapters_http_inventory.LotResponse'
        type: array
      shortfalls:
        items:
          $ref: '#/definitions/internal_adapters_http_inventory.ShortfallResponse'
        type: array
    type: object
  internal_adapters_http_inventory.ItemResponse:
    properties:
      category:
        type: string
      custodian_user_id:
        type: integer
      default_vendor:
        type: string
      id:
        type: string
      location_id:
        type: string
      manufacturer:
        type: string
      max:
        type: integer
      min:
        type: integer
      name:
        type: string
      pct:
        type: integer
      quantity:
        type: integer
      status:
        type: string
      unit:
        type: string
      vendor_id:
        type: string
    type: object
  internal_adapters_http_inventory.LotResponse:
    properties:
      expire_date:
        type: string
      id:
        type: string
      item_id:
        type: string
      lot_no:
        type: string
      quantity:
        type: integer
    type: object
  internal_adapters_http_inventory.ReceiveStockRequest:
    properties:
      expire_date:
        type: string
      lot_no:
        type: string
      quantity:
        type: integer
    required:
    - lot_no
    type: object
  internal_adapters_http_inventory.ReceiveStockResponse:
    properties:
      item:
        $ref: '#/definitions/internal_adapters_http_inventory.ItemResponse'
      lot:
        $ref: '#/definitions/internal_adapters_http_inventory.LotResponse'
    type: object
  internal_adapters_http_inventory.ReorderRequest:
    properties:
      vendor:
        type: string
    required:
    - vendor
    type: object
  internal_adapters_http_inventory.ShortfallResponse:
    properties:
      available:
        type: integer
      lot_id:
        type: string
      lot_no:
        type: string
      requested:
        type: integer
    type: object
  internal_adapters_http_inventory.UpdateDefaultVendorRequest:
    properties:
      vendor:
        type: string
    required:
    - vendor
    type: object
  internal_adapters_http_inventory.UpdateItemRequest:
    properties:
      category:
        type: string
      custodian_user_id:
        type: integer
      location_id:
        type: string
      manufacturer:
        type: string
      max:
        type: integer
      min:
        type: integer
      name:
        type: string
      unit:
        type: string
      vendor_id:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API สำหรับ Thanes LIMS backend (Auth, Sample/Chain-of-Custody,
    Test Result, Equipment, Inventory/Purchase Order, Document, Environment, Notification,
    Audit)
  title: Thanes LIMS Backend API
  version: "1.0"
paths:
...
  /inventory:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_http_inventory.ItemResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: รายการวัสดุคงคลังทั้งหมด
      tags:
      - inventory
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลรายการ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_inventory.CreateItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_inventory.ItemResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: เพิ่มรายการวัสดุคงคลังใหม่
      tags:
      - inventory
  /inventory/{id}:
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_inventory.ItemResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: ดึงข้อมูลรายการวัสดุคงคลังตาม ID
      tags:
      - inventory
    patch:
      consumes:
      - application/json
      description: แก้ไขบางส่วน (partial) — จำนวนคงคลังมาจากการรับล็อต (/receive),
        ผู้ขายเริ่มต้นแก้ผ่าน /default-vendor
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลที่แก้ไข
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_inventory.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_inventory.ItemResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: แก้ไขข้อมูลรายการวัสดุคงคลัง
      tags:
      - inventory
  /inventory/{id}/default-vendor:
    patch:
      consumes:
      - application/json
      description: ใช้เป็นผู้ขายเมื่อระบบสั่งซื้อเพิ่มอัตโนมัติ (auto-reorder)
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: ผู้ขายเริ่มต้น
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_inventory.UpdateDefaultVendorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_inventory.ItemResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: ตั้งค่าผู้ขายเริ่มต้นของรายการ
      tags:
      - inventory
  /inventory/{id}/issue:
    post:
      consumes:
      - application/json
      description: เบิกออกตามล็อตที่ผู้ใช้เลือกเอง (ไม่ใช่ FEFO อัตโนมัติ) — ถ้าจำนวนที่เบิกเกินยอดคงเหลือของล็อต
        ระบบจะไม่หักและตอบยอดคงเหลือจริงกลับมา (applied=false) ให้ระบุล็อตเพิ่มหรือส่งใหม่พร้อม
        force=true เพื่อยอมให้ยอดล็อตติดลบ (ADR 0008)
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: รายการล็อตที่เบิกออก
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_inventory.IssueStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_inventory.IssueStockResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: เบิกของออกจากคลัง
      tags:
      - inventory
  /inventory/{id}/lots:
    get:
      description: ทุกล็อตของรายการ — ใช้เลือกล็อตตอนเบิกออก (Stock Issue)
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_http_inventory.LotResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: รายการล็อตของวัสดุคงคลัง
      tags:
      - inventory
  /inventory/{id}/receive:
    post:
      consumes:
      - application/json
      description: รับสินค้าเข้าคลังตามล็อต — ถ้าเลขล็อตซ้ำกับล็อตเดิมของรายการจะบวกจำนวนเข้าล็อตนั้น
        ไม่งั้นสร้างล็อตใหม่ จำนวนคงคลังของรายการคือผลรวมของทุกล็อต
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลล็อตที่รับเข้า
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_inventory.ReceiveStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_inventory.ReceiveStockResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: รับของเข้าคลัง (สร้าง/เพิ่มล็อต)
      tags:
      - inventory
  /inventory/{id}/reorder:
    post:
      consumes:
      - application/json
      description: สร้าง purchase order จากรายการที่ต่ำกว่า min ด้วยตนเอง
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: ผู้ขาย
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_inventory.ReorderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: สั่งซื้อเพิ่ม (สร้างใบสั่งซื้อ)
      tags:
      - inventory
...
securityDefinitions:
  BearerAuth:
    description: พิมพ์ "Bearer" ตามด้วยเว้นวรรคและ JWT access token เช่น "Bearer eyJhbGciOi..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"